+++ src/commonMain/kotlin/com/github/netomi/sudoku/model/bitsets.kt
@@ -43,6 +43,8 @@ interface SimpleBitSet
     fun nextSetBit(startBit: Int): Int
     fun previousSetBit(startBit: Int): Int
 
+    fun otherSetBit(bit: Int): Int
+
     fun setBits(): Sequence<Int>
     fun setBits(startBit: Int): Sequence<Int>
 
@@ -53,7 +55,7 @@ interface SimpleBitSet
     fun toArray(): IntArray
 }
 
-interface MutableBitSet<in T, out R> : SimpleBitSet
+interface MutableBitSet<in T : SimpleBitSet, out R : MutableBitSet<T, R>> : SimpleBitSet
 {
     fun setAll()
     fun clearAll()
@@ -61,14 +63,15 @@ interface MutableBitSet<in T, out R> : SimpleBitSet
     fun set(bit: Int)
     fun clear(bit: Int)
 
-    fun and(other: T)
-    fun or(other: T)
-    fun andNot(other: T)
+    fun and(other: T): R
+    fun or(other: T): R
+    fun andNot(other: T): R
 
     fun copy(): R
 }
 
-abstract class AbstractBitSetImpl<in T : SimpleBitSet, out R>(final override val size: Int) : MutableBitSet<T, R>
+abstract class AbstractBitSetImpl<in T : SimpleBitSet, out R : MutableBitSet<T, R>>(final override val size: Int)
+    : MutableBitSet<T, R>
 {
     override val bits = BitSet(size)
 
@@ -140,6 +143,16 @@ abstract class AbstractBitSetImpl<in T : SimpleBitSet, out R>(final override val
         return bits.previousSetBit(startBit)
     }
 
+    override fun otherSetBit(bit: Int): Int {
+        require(isBiValue)
+
+        var otherBit = nextSetBit(bit + 1)
+        if (otherBit == -1) {
+            otherBit = previousSetBit(bit - 1)
+        }
+        return otherBit
+    }
+
     override fun setBits(): Sequence<Int> {
         return Sequence { BitIterator(offset, size, false) }
     }
@@ -156,16 +169,19 @@ abstract class AbstractBitSetImpl<in T : SimpleBitSet, out R>(final override val
         return Sequence { BitIterator(startBit, size, true) }
     }
 
-    override fun and(other: T) {
+    override fun and(other: T): R {
         bits.and(other.bits)
+        return this as R
     }
 
-    override fun or(other: T) {
+    override fun or(other: T): R {
         bits.or(other.bits)
+        return this as R
     }
 
-    override fun andNot(other: T) {
+    override fun andNot(other: T): R {
         bits.andNot(other.bits)
+        return this as R
     }
 
     override fun toCollection(): Collection<Int> {
@@ -297,10 +313,6 @@ class MutableValueSet : AbstractBitSetImpl<ValueSet, MutableValueSet>, ValueSet
         return this
     }
 
-    override fun hashCode(): Int {
-        return super.hashCode()
-    }
-
     override fun equals(other: Any?): Boolean {
         if (this === other) return true
         if (other !is MutableValueSet) return false
@@ -329,11 +341,41 @@ class MutableValueSet : AbstractBitSetImpl<ValueSet, MutableValueSet>, ValueSet
     }
 }
 
-class MutableHouseSet : AbstractBitSetImpl<MutableHouseSet, MutableHouseSet>
+interface HouseSet : SimpleBitSet
 {
-    private constructor(size: Int) : super(size)
+    val type: HouseType
 
-    private constructor(other: MutableHouseSet) : super(other.size) {
+    fun houses(grid: Grid): Sequence<House> {
+        return setBits().map { houseIndex ->
+            when(type) {
+                HouseType.ROW    -> grid.getRow(houseIndex)
+                HouseType.COLUMN -> grid.getColumn(houseIndex)
+                HouseType.BLOCK  -> grid.getBlock(houseIndex)
+            }
+        }
+    }
+
+    fun copy(): HouseSet {
+        return toMutableHouseSet()
+    }
+
+    fun toMutableHouseSet(): MutableHouseSet {
+        return MutableHouseSet(this)
+    }
+}
+
+class MutableHouseSet : AbstractBitSetImpl<HouseSet, MutableHouseSet>, HouseSet
+{
+    private var _type: HouseType
+    override val type: HouseType
+        get() = _type
+
+    private constructor(size: Int, type: HouseType) : super(size) {
+        _type = type
+    }
+
+    internal constructor(other: HouseSet) : super(other.size) {
+        _type = other.type
         or(other)
     }
 
@@ -345,12 +387,21 @@ class MutableHouseSet : AbstractBitSetImpl<MutableHouseSet, MutableHouseSet>
         if (this === other) return true
         if (other !is MutableHouseSet) return false
         if (!super.equals(other)) return false
+
+        if (_type != other._type) return false
+
         return true
     }
 
+    override fun hashCode(): Int {
+        var result = super.hashCode()
+        result = 31 * result + _type.hashCode()
+        return result
+    }
+
     companion object {
-        fun empty(grid: Grid): MutableHouseSet {
-            return MutableHouseSet(grid.gridSize)
+        fun empty(grid: Grid, type: HouseType): MutableHouseSet {
+            return MutableHouseSet(grid.gridSize, type)
         }
     }
 }
@@ -365,8 +416,8 @@ interface CellSet : SimpleBitSet
         return setBits(cell.cellIndex + 1).map { cellIndex -> grid.getCell(cellIndex) }
     }
 
-    fun toRowSet(grid: Grid): MutableHouseSet {
-        val rows = MutableHouseSet.empty(grid)
+    fun toRowSet(grid: Grid): HouseSet {
+        val rows = MutableHouseSet.empty(grid, HouseType.ROW)
         cells(grid).forEach { cell -> rows.set(cell.rowIndex) }
         return rows
     }
@@ -383,8 +434,8 @@ interface CellSet : SimpleBitSet
         return if (rows.cardinality() == 1) grid.getRow(rows.firstSetBit()) else null
     }
 
-    fun toColumnSet(grid: Grid): MutableHouseSet {
-        val columns = MutableHouseSet.empty(grid)
+    fun toColumnSet(grid: Grid): HouseSet {
+        val columns = MutableHouseSet.empty(grid, HouseType.COLUMN)
         cells(grid).forEach { cell -> columns.set(cell.columnIndex) }
         return columns
     }
@@ -401,8 +452,8 @@ interface CellSet : SimpleBitSet
         return if (columns.cardinality() == 1) grid.getColumn(columns.firstSetBit()) else null
     }
 
-    fun toBlockSet(grid: Grid): MutableHouseSet {
-        val blocks = MutableHouseSet.empty(grid)
+    fun toBlockSet(grid: Grid): HouseSet {
+        val blocks = MutableHouseSet.empty(grid, HouseType.BLOCK)
         cells(grid).forEach { cell -> blocks.set(cell.blockIndex) }
         return blocks
     }