+++ src/commonMain/kotlin/com/github/netomi/sudoku/solver/techniques/intersectionfinders.kt
@@ -163,8 +163,7 @@ class LockedTripleFinder : BaseHintFinder {
     {
         if (level > subSetSize) return false
 
-        val allVisitedValues = visitedValues.copy()
-        allVisitedValues.or(currentCell.possibleValueSet)
+        val allVisitedValues = visitedValues.copy().or(currentCell.possibleValueSet)
         if (allVisitedValues.cardinality() > subSetSize) return false
 
         visitedCells.set(currentCell.cellIndex)