+++ CHANGELOG.md
@@ -5,9 +5,10 @@
 ### Added
 - added additional hint solver techniques
   - wings (xy-wing, xyz-wing, w-wing)
-
+  - single digit patterns (empty rectangle)
+ 
 ### Fixed
-  
+ 
 
 ## [0.1] - 2020-09-15